// web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        TrackOut: {
            /** Id */
            id: string;
            /** Title */
            title?: string | null;
            /** Mix */
            mix?: string | null;
            /** Artist */
            artist?: string | null;
            /** Original Artist */
            original_artist?: string | null;
            /** Remixer */
            remixer?: string | null;
            /** Composer */
            composer?: string | null;
            /** Album */
            album?: string | null;
            /** Genre */
            genre?: string | null;
            /** Label */
            label?: string | null;
            /** Key Name */
            key_name?: string | null;
            /** Camelot */
            camelot?: string | null;
            /** Bpm */
            bpm?: number | null;
            /** Length S */
            length_s?: number | null;
            /** Rating */
            rating?: number | null;
            /** Color */
            color?: string | null;
            /** Comment */
            comment?: string | null;
            /** Play Count */
            play_count?: number | null;
            /** File Path */
            file_path?: string | null;
            /** File Name */
            file_name?: string | null;
            /** File Type */
            file_type?: number | null;
            /** Bitrate */
            bitrate?: number | null;
            /** Sample Rate */
            sample_rate?: number | null;
            /** Release Year */
            release_year?: number | null;
            /** Release Date */
            release_date?: string | null;
            /** Date Added */
            date_added?: string | null;
            /** Artwork Hash */
            artwork_hash?: string | null;
            /**
             * Source
             * @default local
             */
            source: string;
            /**
             * App Play Count
             * @default 0
             */
            app_play_count: number;
            /** Updated At */
            updated_at?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}